+++ repology_client/__init__.py
@@ -5,8 +5,6 @@
 Asynchronous wrapper for Repology API.
 """
 
-from types import SimpleNamespace
-
 from repology_client._client.v1 import (
     api,
     get_packages,
@@ -16,16 +14,6 @@ from repology_client._client.v1 import (
 from repology_client._client.tools import (
     resolve_package,
 )
-from repology_client._client.experimental import (
-    api as api_exp,
-    distromap,
-)
-
-#: Asynchronous wrapper for experimental endpoints of Repology API.
-exp = SimpleNamespace()
-
-exp.api = api_exp
-exp.distromap = distromap
 
 __all__ = [
     "api",
@@ -33,6 +21,4 @@ __all__ = [
     "get_problems",
     "get_projects",
     "resolve_package",
-
-    "exp",
 ]